Refactor: Organize application structure using Flask Blueprints
