+++ db.py
@@ -1,0 +1,11 @@
+from mysql.connector import pooling
+import os
+
+dbconfig = {
+    'host': 'localhost',
+    'user': 'root',
+    'password': os.getenv('DB_PASSWORD'),
+    'database': 'TravelEcommerceDB'
+}
+
+cnxpool = pooling.MySQLConnectionPool(pool_name="mypool", pool_size=5, **dbconfig)
